test(selectors): renderHHook 에서 필요한 데이터를 미리 set 하도록 변경
